move types into src and fix references
