+++ config/formats/typescriptExportDefinition.ts
@@ -5,7 +5,7 @@ import path = require('path')
 import {treeWalker} from '../utilities/treeWalker'
 
 import type {FormatterArguments} from 'style-dictionary/types/Format'
-import type {w3cTransformedToken} from '~/types/w3cTransformedToken'
+import type {w3cTransformedToken} from '~/src/types/w3cTransformedToken'
 import {prefixTokens} from './utilities/prefixTokens'
 
 const {fileHeader} = StyleDictionary.formatHelpers
