+++ config/platforms/json.ts
@@ -1,5 +1,5 @@
 import type StyleDictionary from 'style-dictionary'
-import type {PlatformInitializer} from '~/types/PlatformInitializer'
+import type {PlatformInitializer} from '~/src/types/PlatformInitializer'
 import {isSource} from '~/config/filters'
 
 export const json: PlatformInitializer = (outputFile, prefix, buildPath): StyleDictionary.Platform => ({
