+++ config/transformers/typographyToCss.ts
@@ -1,6 +1,6 @@
 import type StyleDictionary from 'style-dictionary'
 import {isTypography} from '~/config/filters'
-import type {TypographyTokenValue} from '../../types/TypographyTokenValue'
+import type {TypographyTokenValue} from '~/src/types/TypographyTokenValue'
 import {checkRequiredTokenProperties} from './utilities/checkRequiredTokenProperties'
 import {parseFontFamily} from './fontFamilyToCss'
 import {parseFontWeight} from './fontWeightToNumber'
